requirements.txt
#
# This file is autogenerated by pip-compile with python 3.9
# To update, run:
#
#    pip-compile pyproject.toml
#
certifi==2022.6.15
    # via requests
charset-normalizer==2.1.0
    # via requests
idna==3.3
    # via requests
python-dotenv==0.20.0
    # via rajaongkir-py (pyproject.toml)
requests==2.28.1
    # via rajaongkir-py (pyproject.toml)
urllib3==1.26.11
    # via requests
